#56 :bug: allow local monorepo dependencies to be excluded from requirements.txt easily
